Create retrying client.
Why:

To acknowledge that servers go down or we over-request the API, we need
a client that knows when to sleep for a while.

NOTE: This DOES NOT create a client that will stop for the day when it
knows it's over its query limit -- it will retry for a while and then
exit, so we should figure out a way to throttle our requests so it
doesn't kill all our containers. Though if it does, we can just restart
them. But I'd rather not have to think about it.

This change addresses the need by:

* Having the main client raise an error when it encounters a top-level
error message in the results.

* Having a retrying client catch that error and retry 3 times, sleeping
for longer periods in between requests.