Test #durations convenience method
Why:

Mapping the duration estimates that get returned by the distance matrix
client to something usable is complex process, and we don't want it to
clutter the readability of the main task.

This change addresses the need by:

* Adding a #durations method which returns the 'duration' node of the
response. This method handles cases where duration is unavailable.

* Altering syntax in other areas to express desired syntax, though most
of these changes are untested.

* We also reorder the methods in DistanceMatrixClient to make more
narrative sense.