Add claiming to Waitlist; make Claimable module
Why:

Matt Gardner (@allthesignals) pointed out that it would be safer to mark
ids on the Waitlist as claimed, instead of deleting them outright. It
might be easier to debug later on.

This means that ApiKeys and Waitlist instances are all claimable, and
should have claim!, release!, etc. methods and scopes.

This change addresses the need by:

* Moving claiming logic from ApiKey into a Claimable module

* Including Claimable in ApiKey and Waitlist

* Adding a :claimed boolean to Waitlist