Add a database setup task.
Why:

In order to set up the central Waitlist database for workers to work
off, we needed to have a setup task that builds the Waitlist from
distinct target_ids in the TravelTimes table.

This change addresses the need by:

* Creating a task that locks the Waitlist, empties it, fills it with
distict origin IDs from the TravelTime target_id, and sets :claimed to
false on every one of them.

* Adds the setup task to the rake tasks.

* Adds a statistics task to the rake tasks to tell whether the table is
fully built and provide information on task progress.

* Fixes a bug found in Claimable during this process.