'Fix' SSL errors.
Why:

The Net::HTTP library is throwing SSL connection errors when trying to
request the Google API, and the Google API is ONLY accepting requests
over HTTPS. Instead of trying to install certificate bundles, which I
admittedly do not understand, we set OpenSSL to not verify the recipient
of the request.

This change addresses the need by:

* Setting SSL to not verify the 'peer', to avoid the connection error.
This is an insecure fix, but security is not a big concern for this part
of the project.

* Removing HTTParty from Gemfile and client, and replacing it with
Net::HTTP.