Prepares task for run; updates reference IDs.
Why:

The RunTask wasn't working, and was returning unexpected values, because
I was initializing the RetryingClient task incorrectly. I assumed that
it would inherit directly from DistanceMatrixClient, but it needed
DistanceMatrixClient as a parameter to delegate to.

Additionally, all of the querying for origins was based on
target_id, when in fact it should be input_id; target_id is the
destination.

This change addresses the need by:

* Update the initialization of the RetryingClient.

* Add rescue clauses to ensure resources are released.

* Change target_id to input_id where relevant.