Tweak for production.
Why:

The first request made was a 'Bad Request', with the message from Google
saying "That's all we know." Upon some further investigation, we found
that Google limits URLs to about 2000 characters.

This change addresses the need by:

* Rounding coordinates to 4 decimal places. Four decimal places is
roughly equivalent 10m in each direction in this region. Rounding keeps
us close to but not exactly at the center of each grid cell. For a
request of 100 destinations, it reduces the request size by 400-500
characters.

* Limiting calls batches of 50. The API is limited based on the number
of elements returned, not the number of calls, so this doesn't hurt us
in terms of API limits.